Fix empty string bug.
In cases where a part of a message is empty (such as Message-ID), the
string would be None, and would cause the script to crash out. This
is fixed by adding a function that returns an empty string when passed
a None object, or the string itself it is a valid string.